Refactor to single test with nested variants
